+++ packages/core/hooks/src/use-animate-presence/use-animate-presence.ts
@@ -9,18 +9,22 @@ export function useAnimatePresence(props: UseVisibilityProps) {
   const { isOpen, duration = 500 } = props
 
   const [isVisible, setIsVisible] = useState(false)
+  const [isClosing, setIsClosing] = useState(false)
 
   useLayoutEffect(() => {
     if (isOpen) {
       setIsVisible(true)
+      setIsClosing(false)
     }
     else {
+      setIsClosing(true)
       const timer = setTimeout(() => {
         setIsVisible(false)
+        setIsClosing(false)
       }, duration)
       return () => clearTimeout(timer)
     }
   }, [isOpen, duration])
 
-  return { isVisible }
+  return { isVisible, isClosing }
 }