+++ packages/core/ui/src/index.ts
@@ -1,5 +1,7 @@
 export * from '@srcube-taro/button'
 
+export * from '@srcube-taro/dialog'
+
 export * from '@srcube-taro/input'
 
 export * from '@srcube-taro/layouts'