+++ packages/ui/dialog/src/dialog-body.tsx
@@ -1,0 +1,10 @@
+import type { ModalBodyProps } from '@srcube-taro/modal'
+import { ModalBody } from '@srcube-taro/modal'
+
+export interface DialogBodyProps extends ModalBodyProps {}
+
+const DialogBody = ModalBody
+
+DialogBody.displayName = 'Srcube.DialogBody'
+
+export default DialogBody