+++ packages/ui/dialog/src/dialog-footer.tsx
@@ -1,0 +1,10 @@
+import type { ModalFooterProps } from '@srcube-taro/modal'
+import { ModalFooter } from '@srcube-taro/modal'
+
+export interface DialogFooterProps extends ModalFooterProps {}
+
+const DialogFooter = ModalFooter
+
+DialogFooter.displayName = 'Srcube.DialogFooter'
+
+export default DialogFooter