+++ packages/ui/dialog/src/index.ts
@@ -1,0 +1,11 @@
+export type { DialogProps } from './dialog'
+export { default as Dialog } from './dialog'
+
+export type { DialogBodyProps } from './dialog-body'
+export { default as DialogBody } from './dialog-body'
+
+export type { DialogFooterProps } from './dialog-footer'
+export { default as DialogFooter } from './dialog-footer'
+
+export type { DialogHeaderProps } from './dialog-header'
+export { default as DialogHeader } from './dialog-header'