+++ packages/ui/modal/src/index.ts
@@ -3,12 +3,16 @@ export { default as Modal } from './modal'
 
 export type { ModalBodyProps } from './modal-body'
 export { default as ModalBody } from './modal-body'
-export type { ModalContentProps } from './modal-content'
-export { default as ModalContent } from './modal-content'
 export type { ModalFooterProps } from './modal-footer'
 export { default as ModalFooter } from './modal-footer'
 export type { ModalHeaderProps } from './modal-header'
 export { default as ModalHeader } from './modal-header'
+
+// context
+export * from './context'
+
+// component hook
+export * from './use'
 
 // hooks
 export { useDisclosure } from '@srcube-taro/hooks'