+++ packages/core/hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @srcube-taro/hooks
 
+## 0.0.2
+
+### Patch Changes
+
+- Components and hooks, utils package
+  dialog
+
 ## 0.0.1
 
 ### Patch Changes