+++ packages/core/ui/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @srcube-taro/ui
 
+## 0.0.5
+
+### Patch Changes
+
+- Components and hooks, utils package
+  dialog
+- Updated dependencies
+  - @srcube-taro/theme@0.0.5
+  - @srcube-taro/button@0.0.5
+  - @srcube-taro/modal@0.0.2
+  - @srcube-taro/dialog@0.0.1
+  - @srcube-taro/input@0.0.3
+  - @srcube-taro/layouts@0.0.5
+  - @srcube-taro/spinner@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes