+++ packages/ui/modal/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @srcube-taro/modal
 
+## 0.0.2
+
+### Patch Changes
+
+- Components and hooks, utils package
+  dialog
+- Updated dependencies
+  - @srcube-taro/hooks@0.0.2
+  - @srcube-taro/theme@0.0.5
+
 ## 0.0.1
 
 ### Patch Changes